refactor(domain): domain package를 만듬
util보다는 domain에 있는 것이 의미상 더 맞기에 RacingCarGame, RacingCarGameInput, RacingCarGameOutput 관련 Test들을 domain으로 분류했습니다.
